+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Display the numbers of registered items below the lists in the syntax editor.
 - Deprecate the “Make Snake Case,” “Make Camel Case,” and “Make Pascal Case” commands in the Text > Transformations menu.
 - [dev] Migrate the input fields in the Find window to SwiftUI.
+- [non-AppStore ver.] Update Sparkle from 2.7.1 to 2.8.0-beta.2.
 
 
 ### Fixes