+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Deprecate the “Make Snake Case,” “Make Camel Case,” and “Make Pascal Case” commands in the Text > Transformations menu.
 - [dev] Migrate the input fields in the Find window to SwiftUI.
 - [non-AppStore ver.] Update Sparkle from 2.7.1 to 2.8.0-beta.3.
+- [beta][dev] Update the build environment to Xcode 26.0 RC.
 
 
 ### Fixes